+++ pkg/provider/surge.go
@@ -46,6 +46,8 @@ func checkSurgeSupport(p proxy.Proxy) bool {
 		}
 	case *proxy.Trojan:
 		return true
+	case *proxy.Vless:
+		return true
 	default:
 		return false
 	}